+++ stac_fastapi/core/stac_fastapi/core/core.py
@@ -312,22 +312,14 @@ class CoreClient(AsyncBaseCoreClient):
             self.item_serializer.db_to_stac(item, base_url=base_url) for item in items
         ]
 
-        context_obj = None
-        if self.extension_is_enabled("ContextExtension"):
-            context_obj = {
-                "returned": len(items),
-                "limit": limit,
-            }
-            if maybe_count is not None:
-                context_obj["matched"] = maybe_count
-
         links = await PagingLinks(request=request, next=next_token).get_links()
 
         return stac_types.ItemCollection(
             type="FeatureCollection",
             features=items,
             links=links,
-            context=context_obj,
+            numReturned=len(items),
+            numMatched=maybe_count,
         )
 
     async def get_item(
@@ -633,22 +625,14 @@ class CoreClient(AsyncBaseCoreClient):
                 for feat in items
             ]
 
-        context_obj = None
-        if self.extension_is_enabled("ContextExtension"):
-            context_obj = {
-                "returned": len(items),
-                "limit": limit,
-            }
-            if maybe_count is not None:
-                context_obj["matched"] = maybe_count
-
         links = await PagingLinks(request=request, next=next_token).get_links()
 
         return stac_types.ItemCollection(
             type="FeatureCollection",
             features=items,
             links=links,
-            context=context_obj,
+            numReturned=len(items),
+            numMatched=maybe_count,
         )